+++ coruscant/src/actions/project.js
@@ -26,10 +26,6 @@ export function setProjects(projects) {
   return actionCreator(Constants.SET_PROJECTS, { projects })
 }
 
-export function setUserSearch(user) {
-  return actionCreator(Constants.SET_USER_SEARCH, { user })
-}
-
 export function getUserProjects(token) {
   return async function (dispatch) {
     try {
@@ -80,48 +76,6 @@ export function getBuildStatuses(token, id, limit) {
       })
       if (result) {
         dispatch(setBuildStatuses(result))
-        return { result, error }
-      }
-      console.log(error)
-      return { result: null, error }
-    } catch (error) {
-      console.log(error)
-      return { result: null, error }
-    }
-  }
-}
-
-export function addGithubToken(token, body) {
-  return async function () {
-    try {
-      const fetchClient = new FetchClient()
-      fetchClient.setAuthToken(token)
-      const { result, error } = await fetchClient.post({
-        url: '/auth/github',
-        body
-      })
-      if (result) {
-        return { result, error }
-      }
-      console.log(error)
-      return { result: null, error }
-    } catch (error) {
-      console.log(error)
-      return { result: null, error }
-    }
-  }
-}
-
-export function addTravisToken(token, body) {
-  return async function () {
-    try {
-      const fetchClient = new FetchClient()
-      fetchClient.setAuthToken(token)
-      const { result, error } = await fetchClient.post({
-        url: '/auth/travis',
-        body
-      })
-      if (result) {
         return { result, error }
       }
       console.log(error)
