+++ coruscant/src/containers/ProjectOverview.js
@@ -7,10 +7,10 @@ import _ from 'lodash'
 import { getProject } from '../actions/project'
 import ProjectHeader from '../components/ProjectHeader'
 import ProjectDescription from '../components/ProjectDescription'
-import ProjectIssues from '../components/ProjectIssues'
-import ProjectBuilds from '../components/ProjectBuilds'
+import ProjectIssues from './ProjectIssues'
+import ProjectBuilds from './ProjectBuilds'
 import ProjectLinks from '../components/ProjectLinks'
-import ProjectContributors from '../components/ProjectContributors'
+import ProjectContributors from './ProjectContributors'
 import Theme from '../../style/theme'
 import Spinner from './../components/Spinner'
