+++ coruscant/src/reducers/index.js
@@ -3,9 +3,13 @@ import { combineReducers } from 'redux'
 import session from './session'
 import projects from './projects'
 import projectMembers from './projectMembers'
+import projectIssues from './projectIssues'
+import projectBuilds from './projectBuilds'
 
 export default combineReducers({
   session,
   projects,
-  projectMembers
+  projectMembers,
+  projectIssues,
+  projectBuilds
 })
